Update websocket.h
There's a bug in the payload processing if you receive a message larger than the internal buffer size (4K). You need to move the do_read down so that it keeps on reading (with state_ == Payload) until the entire message is processed and remaining_length_ == 0. Currently you only call do_read if reaminging_length_ == 0 on the first read and message fits in the buffer.